Combine example with example from #832
